Rework burton init() function in the direction of getting the proper dual mesh. Add edge vertices, and make wedges using them.
Add function midpoint that takes a const edge_t * and returns a point_t with
with the midpoint. There was some name collisions with this version of midpoint
and the one in point.h. Got around it by using flexi::midpoint() for the latter.

Edit the burton.cc test to print out coordinates of the entities so that
numbering could be inferred. The numbering of edges is not consistent with the
proposal made to nathaniel. comment out the wedges and corners output until
it can be worked out correctly. The test fails as the blessed file was not
updated.