state -> data overhaul
Part of this change is cosmetic: state_t has been renamed to data_t

Part of the change is real: data_t is now a singleton data manager.  One
consequence of this is that some of the tests recreate the mesh several
times from within a single translation unit.  This causes data to
collide for internal mesh quantities (namely coordinates).  The
temporary fix is to clear the data inside of init_parameters.  This is
probably not a good long-term solution and we will have to figure out
how to make the data manager more robust (simple way is to punt this to
the developer of the specialization).