Reordered macros to avoid variadic macro warning
This mostly just affected the register_data macro. The versions
argument was moved to the end of the interface because it is used
by all storage types. This means that it can be treated as a variadic
argument so that there is at least one varidic argument for each
storage type.