Fixing Legion unit tests.
This commit adds a call to 'gtest_init', a user-defined function (from
the point of view of cinch) that is called during the gtest policy
initialization. For Legion unit tests, this function is responsible
for registering all Legion tasks with the Legion runtime. Also required
is a driver function definition.

It seems like we may be able to define a more elegant solution, but
this should work for the time being. The primary downside to the
current approach is that it still requires an empty translation unit
for the test to compile. We should think about a better solution.