This could break things...
I'm trying to remove all references to the serial runtime. If you want
to run serial now, you should just build with MPI and execute 1 rank
(I have not really verified that this works.)

I have tested this commit with both the Legion and MPI backends, and all
of the tests pass.

Also, I have added a 'developer' directory that contains some
documentation for FleCSI developers.