Runtime Flags
This commit resovles issue #7 on gitlab.lanl.gov:laristra/flecsi.git.

Unrecognized flags are passed to the backend. FleCSI provides
a boost::program_options interface for specialization and application
developers to add and use command-line options.

This commit also finishes the refactor of the runtime interface into the
more standard initialize/finalize style. Initial testing seems to
provide the functionality that we want.