Refactor of Gitlab CI
This commit restructures the CI to use customizable containers from the
gitlab-ci branch of flecsi. Dependencies are managed with spack
environments. Additionally, it splits the build steps for dependencies
such that packages that are unlikely to be modified by the commit are
handled as part of an 'environment' that can be setup in the gitlab-ci
branch. The most commonly modified package (namely legion) is built
after the commit spack repo has been added. This still allows
customization of other packages. However, if customization is not
needed, the build will be faster.

- Split up monolithic gitlab yaml file (subfiles are in .gitlab directory)
- Updated formatting logic with CMake FORMAT_ONLY option, so that the
  flecsi formatting target can be executed without full configuration
  (and dependencies).
- Added basic reporting mechanism to display relevant information during
  build (useful for debugging).
- Added basic tutorial unit tests.

TODO:

- Still need to flesh out the full matrix of tests that should be run.
- Need to add bare-metal testing (optional for devel branch)